MdeModulePkg: Introduce and use PcdMemoryAttributesTableVersion
Linux kernel supports the memory attributes table v2 since version 6.2.
However, UEFI can be used with older kernel versions that are still
maintainable. So instead of relying on a macro that defines the memory
attributes table, a corresponding PCD is introduced to decide which
version of the table UEFI should create.

Signed-off-by: Alexander Goncharov <chat@joursoir.net>